+++ src/Client/ClientCore/Systems/Input/MouseButton.cs
@@ -16,7 +16,28 @@
 
 namespace Sovereign.ClientCore.Systems.Input;
 
+/// <summary>
+///     Mouse button enumeration.
+/// </summary>
 public enum MouseButton
 {
-    
+    /// <summary>
+    ///     Left mouse button.
+    /// </summary>
+    Left,
+
+    /// <summary>
+    ///     Middle mouse button.
+    /// </summary>
+    Middle,
+
+    /// <summary>
+    ///     Right mouse button.
+    /// </summary>
+    Right,
+
+    /// <summary>
+    ///     Other mouse button.
+    /// </summary>
+    Other
 }